2nd refactor
