+++ app/api/v2/admin/logs/sign-ups/route.ts
