+++ app/api/v2/books/available/route.ts
@@ -1,0 +1,15 @@
+import prisma from "@/prisma/client";
+import { NextRequest, NextResponse } from "next/server";
+
+// fetches all the book that are rentable
+export async function GET(request : NextRequest) {
+  const books = await prisma.storeBooks.findMany({
+    where : {rentable : true},
+    select : {
+      id : true , 
+      title : true
+    }
+  })
+
+  return NextResponse.json(books)
+}
